feature : backgroud image 최적화
- 기존 이미지 사이즈가 너무커 압축하여 최적화 작업함 ( 데스크탑용하고 모바일용으로 나눠 반응형 작업, webp로 확장자 변경)
